Make checkboxes consistent with XAML API
Checkboxes need a checked and unchecked event so that you can do dynamic UI (ui that enables or disables when you check the box),

Also, the naming was inconsistent with XAML, where IsChecked is the property and Checked and Unchecked are events.  This change aligns with the XAML api which will reduce cognitive load for devs moving back and forth between the two.