Fix json serialization of boolean values
The JValue class was serializing booleans as 'True' and 'False' when they should be written out as 'true' and 'false'.

Changed the serialization of booleans to output the correct case. The spec for json is very particular on this point. Booleans must be lowercase.

This is NOT a breaking change for existing programs. The deserializer will accept either upper or lower case, but now it will emit the correct case.