Reassignment of values should not throw
Assignment of properties within an object was using .Add()
on the HashTable, which will throw if the value already exists.

This makes it hard to read a json blob and then change values before writing it back out (e.g., in the scenario of using json/bson for config).

This change uses subscript assignment instead, which will not throw if the value exists already. This version will just overwrite the existing value. If you want to ensure that you are not overwrtiting a value, use the Contains() test first.