Polymorphic output with enhanced factory for input
On output, the type of the actual instance is used for
property and field enumeration instead of the return
type of the property. This ensures that all of the members
of the actual instance are output.

On input, the definition of 'InstanceFactory' is expanded
to provide more context for the instance factory to use
when creating instances to be populated from the json.